More minor changes, Updated create post component, url path crash prevention, and conditionally buttons that show depending on who is viewing a profile page
In this update, I updated the create post component to render the file image or preview a video file before creating a post. I also added a prevention that would catch an error from crashing the application if a user tried to go to a profile/user path that did not exist. Lastly, I changed the edit profile button and create post button to be conditionally rendered on a user profile page so that it would only appear if the user was viewing their own profile